requirements.txt
django
django-crispy-forms
django-filter
django-widget-tweaks
django-select2
django-redis

